+++ package.json
@@ -14,15 +14,21 @@
         ]
     },
     "dependencies": {
+        "@babel/core": "^7.3.4",
+        "@babel/plugin-transform-react-jsx": "^7.3.0",
+        "@babel/preset-env": "^7.3.4",
         "@emotion/babel-preset-css-prop": "^10.0.7",
         "@emotion/core": "^10.0.7",
         "@emotion/styled": "^10.0.7",
+        "babel-loader": "^8.0.5",
+        "babel-plugin-emotion": "^10.0.7",
+        "clean-webpack-plugin": "^1.0.1",
         "axios": "^0.18.0",
         "cross-env": "^5.2.0",
         "debug": "~2.6.9",
         "express": "~4.16.0",
         "gm": "^1.23.1",
-        "htm": "^2.0.0",
+        "html-webpack-plugin": "^3.2.0",
         "image-size": "^0.7.1",
         "media-platform-js-sdk": "^6.0.5",
         "mkdir-recursive": "^0.4.0",
@@ -31,29 +37,22 @@
         "react": "^16.8.1",
         "react-dom": "^16.8.1",
         "react-transition-group": "^2.5.3",
-        "sharp": "^0.21.2"
+        "sharp": "^0.21.2",
+        "terser-webpack-plugin": "^1.2.2",
+        "webpack": "^4.29.3"
     },
     "devDependencies": {
-        "@babel/core": "^7.3.4",
-        "@babel/plugin-transform-react-jsx": "^7.3.0",
-        "@babel/preset-env": "^7.3.4",
         "babel-jest": "^24.1.0",
-        "babel-loader": "^8.0.5",
-        "babel-plugin-emotion": "^10.0.7",
         "babel-polyfill": "^6.26.0",
-        "clean-webpack-plugin": "^1.0.1",
         "enzyme": "^3.9.0",
         "enzyme-adapter-react-16": "^1.10.0",
         "enzyme-to-json": "^3.3.5",
         "eslint": "^5.12.0",
-        "html-webpack-plugin": "^3.2.0",
         "jest": "^24.1.0",
         "jest-emotion": "^10.0.7",
         "jsdom": "^13.2.0",
         "nodemon": "^1.18.9",
         "prettier": "^1.16.4",
-        "terser-webpack-plugin": "^1.2.2",
-        "webpack": "^4.29.3",
         "webpack-cli": "^3.2.3",
         "webpack-dev-middleware": "^3.5.2",
         "webpack-hot-middleware": "^2.24.3"
