+++ src-client/js/components/icons/smallIconCss.js
@@ -1,0 +1,8 @@
+import { css } from '@emotion/core';
+
+const smallIconCss = css`
+    width: 2em;
+    height: 2em;
+`;
+
+export default smallIconCss;
