+++ src-client/js/components/pages/PageWrapper.js
@@ -1,0 +1,12 @@
+import styled from '@emotion/styled';
+
+const PageWrapper = styled.div`
+    margin: 0;
+    padding: 0;
+    height: 30em;
+    width: 20em;
+    display: flex;
+    flex-direction: column;
+`;
+
+export default PageWrapper;
