+++ src-client/js/components/steps/StatusRow.js
@@ -1,0 +1,12 @@
+import styled from '@emotion/styled';
+
+const StatusRow = styled.div`
+    display: flex;
+    align-items: center;
+
+    & > * {
+        margin-left: 0.5em;
+    }
+`;
+
+export default StatusRow;
