+++ test/client/App.test.js
@@ -1,19 +1,19 @@
 import React from 'react';
 import { render, mount } from 'enzyme';
-import UploadForm from '../../src-client/js/components/UploadForm';
+import App from '../../src-client/js/components/App';
 
 const findCollectionInput = wrapper =>
     wrapper.findWhere(
         n => n.name() === 'TextInput' && n.prop('id') === 'collection'
     );
 
-describe('UploadForm', () => {
+describe('App', () => {
     test('matches snapshot', () => {
-        expect(render(<UploadForm />)).toMatchSnapshot();
+        expect(render(<App />)).toMatchSnapshot();
     });
 
     test('does not allow to submit in initial state', () => {
-        const wrapper = mount(<UploadForm />);
+        const wrapper = mount(<App />);
         const submitButton = wrapper.find(
             'button[data-testid="submit-button"]'
         );
@@ -23,7 +23,7 @@ describe('UploadForm', () => {
     });
 
     test('shows error indication when collection name is deleted', () => {
-        const wrapper = mount(<UploadForm />);
+        const wrapper = mount(<App />);
         const collectionInput = findCollectionInput(wrapper);
 
         expect(wrapper.find('[data-testid="collection-status"]')).toHaveLength(
@@ -41,7 +41,7 @@ describe('UploadForm', () => {
     });
 
     test('handles single file selection correctly', () => {
-        const wrapper = mount(<UploadForm />);
+        const wrapper = mount(<App />);
         const uploaderInput = wrapper.find('input#uploader');
 
         expect(wrapper.find('[data-testid="uploader-status"]')).toHaveLength(0);
@@ -60,7 +60,7 @@ describe('UploadForm', () => {
     });
 
     test('handles multipe file selection correctly', () => {
-        const wrapper = mount(<UploadForm />);
+        const wrapper = mount(<App />);
         const uploaderInput = wrapper.find('input#uploader');
 
         expect(wrapper.find('[data-testid="uploader-status"]')).toHaveLength(0);
@@ -84,7 +84,7 @@ describe('UploadForm', () => {
     });
 
     test('handles file selection cancellation correctly', () => {
-        const wrapper = mount(<UploadForm />);
+        const wrapper = mount(<App />);
         const uploaderInput = wrapper.find('input#uploader');
 
         expect(wrapper.find('[data-testid="uploader-status"]')).toHaveLength(0);
@@ -106,7 +106,7 @@ describe('UploadForm', () => {
     });
 
     test('handles file types correctly', () => {
-        const wrapper = mount(<UploadForm />);
+        const wrapper = mount(<App />);
         const uploaderInput = wrapper.find('input#uploader');
 
         expect(wrapper.find('[data-testid="uploader-status"]')).toHaveLength(0);
@@ -139,7 +139,7 @@ describe('UploadForm', () => {
     });
 
     test('enables and disables the submission button correctly', () => {
-        const wrapper = mount(<UploadForm />);
+        const wrapper = mount(<App />);
         const uploaderInput = wrapper.find('input#uploader');
         const collectionInput = findCollectionInput(wrapper);
